requirements.txt
biopython==1.81
datasets==2.15.0
einops==0.7.0
enformer-pytorch==0.8.8
fsspec==2023.10.0
genomic-benchmarks==0.0.9
git-lfs==1.6
h5py==3.10.0
huggingface-hub==0.19.4
hydra-core==1.3.2
ipdb==0.13.13
matplotlib==3.7.4
notebook==7.1.1
nvitop==1.3.2
omegaconf==2.3.0
pandas==2.0.3
pyfaidx==0.8.1.1
pysam==0.22.0
pytest==8.0.2
pytorch-lightning==1.8.6
rich==13.7.0
seaborn==0.13.2
scikit-learn==1.3.2
timm==0.9.16
torch==2.2.0
torchaudio==2.2.0
torchdata==0.7.1
torchmetrics==1.2.1
torchtext==0.17.0
torchvision==0.17.0
tqdm==4.66.1
transformers==4.38.1
triton==2.2.0
wandb==0.13.5
flash-attn==2.5.6
causal-conv1d===1.2.0.post2
mamba-ssm==1.2.0.post1